fix(endpoint): move /auth/profile to /users/profile to be more fitting
