+++ src/main/java/com/github/liuchangming88/ecommerce_backend/api/controller/auth/AuthenticationController.java
@@ -1,12 +1,10 @@
 package com.github.liuchangming88.ecommerce_backend.api.controller.auth;
 
 import com.github.liuchangming88.ecommerce_backend.api.model.*;
-import com.github.liuchangming88.ecommerce_backend.model.LocalUser;
 import com.github.liuchangming88.ecommerce_backend.service.UserService;
 import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -38,19 +36,6 @@ public class AuthenticationController {
     public ResponseEntity<String> verifyUser (@RequestParam String token) {
         userService.verifyUser(token);
         return new ResponseEntity<>("User has been verified!", HttpStatus.OK);
-    }
-
-    @GetMapping(path = "/profile")
-    public ResponseEntity<RegistrationResponse> getProfile (@AuthenticationPrincipal LocalUser localUser) {
-        return new ResponseEntity<>(
-                new RegistrationResponse(
-                        localUser.getUsername(),
-                        localUser.getEmail(),
-                        localUser.getFirstName(),
-                        localUser.getLastName(),
-                        ""
-                ), HttpStatus.OK
-        );
     }
 
     @PostMapping(path = "/forgot-password")
